# Replace SAO with ASSAO as Godot's new SSAO