--- a/scene/resources/environment.cpp
+++ b/scene/resources/environment.cpp
@@ -368,13 +368,40 @@ float Environment::get_ssao_intensity() const {
 	return ssao_intensity;
 }
 
-void Environment::set_ssao_bias(float p_bias) {
-	ssao_bias = p_bias;
+void Environment::set_ssao_power(float p_power) {
+	ssao_power = p_power;
 	_update_ssao();
 }
 
-float Environment::get_ssao_bias() const {
-	return ssao_bias;
+float Environment::get_ssao_power() const {
+	return ssao_power;
+}
+
+void Environment::set_ssao_detail(float p_detail) {
+	ssao_detail = p_detail;
+	_update_ssao();
+}
+
+float Environment::get_ssao_detail() const {
+	return ssao_detail;
+}
+
+void Environment::set_ssao_horizon(float p_horizon) {
+	ssao_horizon = p_horizon;
+	_update_ssao();
+}
+
+float Environment::get_ssao_horizon() const {
+	return ssao_horizon;
+}
+
+void Environment::set_ssao_sharpness(float p_sharpness) {
+	ssao_sharpness = p_sharpness;
+	_update_ssao();
+}
+
+float Environment::get_ssao_sharpness() const {
+	return ssao_sharpness;
 }
 
 void Environment::set_ssao_direct_light_affect(float p_direct_light_affect) {
@@ -395,35 +422,18 @@ float Environment::get_ssao_ao_channel_affect() const {
 	return ssao_ao_channel_affect;
 }
 
-void Environment::set_ssao_blur(SSAOBlur p_blur) {
-	ssao_blur = p_blur;
-	_update_ssao();
-}
-
-Environment::SSAOBlur Environment::get_ssao_blur() const {
-	return ssao_blur;
-}
-
-void Environment::set_ssao_edge_sharpness(float p_edge_sharpness) {
-	ssao_edge_sharpness = p_edge_sharpness;
-	_update_ssao();
-}
-
-float Environment::get_ssao_edge_sharpness() const {
-	return ssao_edge_sharpness;
-}
-
 void Environment::_update_ssao() {
 	RS::get_singleton()->environment_set_ssao(
 			environment,
 			ssao_enabled,
 			ssao_radius,
 			ssao_intensity,
-			ssao_bias,
+			ssao_power,
+			ssao_detail,
+			ssao_horizon,
+			ssao_sharpness,
 			ssao_direct_light_affect,
-			ssao_ao_channel_affect,
-			RS::EnvironmentSSAOBlur(ssao_blur),
-			ssao_edge_sharpness);
+			ssao_ao_channel_affect);
 }
 
 // SDFGI
@@ -1150,26 +1160,29 @@ void Environment::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("get_ssao_radius"), &Environment::get_ssao_radius);
 	ClassDB::bind_method(D_METHOD("set_ssao_intensity", "intensity"), &Environment::set_ssao_intensity);
 	ClassDB::bind_method(D_METHOD("get_ssao_intensity"), &Environment::get_ssao_intensity);
-	ClassDB::bind_method(D_METHOD("set_ssao_bias", "bias"), &Environment::set_ssao_bias);
-	ClassDB::bind_method(D_METHOD("get_ssao_bias"), &Environment::get_ssao_bias);
+	ClassDB::bind_method(D_METHOD("set_ssao_power", "power"), &Environment::set_ssao_power);
+	ClassDB::bind_method(D_METHOD("get_ssao_power"), &Environment::get_ssao_power);
+	ClassDB::bind_method(D_METHOD("set_ssao_detail", "detail"), &Environment::set_ssao_detail);
+	ClassDB::bind_method(D_METHOD("get_ssao_detail"), &Environment::get_ssao_detail);
+	ClassDB::bind_method(D_METHOD("set_ssao_horizon", "horizon"), &Environment::set_ssao_horizon);
+	ClassDB::bind_method(D_METHOD("get_ssao_horizon"), &Environment::get_ssao_horizon);
+	ClassDB::bind_method(D_METHOD("set_ssao_sharpness", "sharpness"), &Environment::set_ssao_sharpness);
+	ClassDB::bind_method(D_METHOD("get_ssao_sharpness"), &Environment::get_ssao_sharpness);
 	ClassDB::bind_method(D_METHOD("set_ssao_direct_light_affect", "amount"), &Environment::set_ssao_direct_light_affect);
 	ClassDB::bind_method(D_METHOD("get_ssao_direct_light_affect"), &Environment::get_ssao_direct_light_affect);
 	ClassDB::bind_method(D_METHOD("set_ssao_ao_channel_affect", "amount"), &Environment::set_ssao_ao_channel_affect);
 	ClassDB::bind_method(D_METHOD("get_ssao_ao_channel_affect"), &Environment::get_ssao_ao_channel_affect);
-	ClassDB::bind_method(D_METHOD("set_ssao_blur", "mode"), &Environment::set_ssao_blur);
-	ClassDB::bind_method(D_METHOD("get_ssao_blur"), &Environment::get_ssao_blur);
-	ClassDB::bind_method(D_METHOD("set_ssao_edge_sharpness", "edge_sharpness"), &Environment::set_ssao_edge_sharpness);
-	ClassDB::bind_method(D_METHOD("get_ssao_edge_sharpness"), &Environment::get_ssao_edge_sharpness);
 
 	ADD_GROUP("SSAO", "ssao_");
 	ADD_PROPERTY(PropertyInfo(Variant::BOOL, "ssao_enabled"), "set_ssao_enabled", "is_ssao_enabled");
-	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_radius", PROPERTY_HINT_RANGE, "0.1,128,0.01"), "set_ssao_radius", "get_ssao_radius");
-	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_intensity", PROPERTY_HINT_RANGE, "0.0,128,0.01"), "set_ssao_intensity", "get_ssao_intensity");
-	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_bias", PROPERTY_HINT_RANGE, "0.001,8,0.001"), "set_ssao_bias", "get_ssao_bias");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_radius", PROPERTY_HINT_RANGE, "0.01,16,0.01,or_greater"), "set_ssao_radius", "get_ssao_radius");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_intensity", PROPERTY_HINT_RANGE, "0,16,0.01,or_greater"), "set_ssao_intensity", "get_ssao_intensity");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_power", PROPERTY_HINT_EXP_EASING), "set_ssao_power", "get_ssao_power");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_detail", PROPERTY_HINT_RANGE, "0,5,0.01"), "set_ssao_detail", "get_ssao_detail");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_horizon", PROPERTY_HINT_RANGE, "0,1,0.01"), "set_ssao_horizon", "get_ssao_horizon");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_sharpness", PROPERTY_HINT_RANGE, "0,1,0.01"), "set_ssao_sharpness", "get_ssao_sharpness");
 	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_light_affect", PROPERTY_HINT_RANGE, "0.00,1,0.01"), "set_ssao_direct_light_affect", "get_ssao_direct_light_affect");
 	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_ao_channel_affect", PROPERTY_HINT_RANGE, "0.00,1,0.01"), "set_ssao_ao_channel_affect", "get_ssao_ao_channel_affect");
-	ADD_PROPERTY(PropertyInfo(Variant::INT, "ssao_blur", PROPERTY_HINT_ENUM, "Disabled,1x1,2x2,3x3"), "set_ssao_blur", "get_ssao_blur");
-	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "ssao_edge_sharpness", PROPERTY_HINT_RANGE, "0,32,0.01"), "set_ssao_edge_sharpness", "get_ssao_edge_sharpness");
 
 	// SDFGI
 
@@ -1366,11 +1379,6 @@ void Environment::_bind_methods() {
 	BIND_ENUM_CONSTANT(GLOW_BLEND_MODE_SOFTLIGHT);
 	BIND_ENUM_CONSTANT(GLOW_BLEND_MODE_REPLACE);
 	BIND_ENUM_CONSTANT(GLOW_BLEND_MODE_MIX);
-
-	BIND_ENUM_CONSTANT(SSAO_BLUR_DISABLED);
-	BIND_ENUM_CONSTANT(SSAO_BLUR_1x1);
-	BIND_ENUM_CONSTANT(SSAO_BLUR_2x2);
-	BIND_ENUM_CONSTANT(SSAO_BLUR_3x3);
 
 	BIND_ENUM_CONSTANT(SDFGI_CASCADES_4);
 	BIND_ENUM_CONSTANT(SDFGI_CASCADES_6);