… doc/classes/Environment.xml
… doc/classes/ProjectSettings.xml
… doc/classes/RenderingServer.xml
… servers/rendering/renderer_rd/effects_rd.cpp
… servers/rendering/renderer_rd/renderer_scene_render_rd.cpp
… servers/rendering/renderer_rd/shaders/SCsub
… servers/rendering/renderer_rd/shaders/ssao.glsl
… servers/rendering/renderer_rd/shaders/ssao_blur.glsl
… servers/rendering/renderer_rd/shaders/ssao_downsample.glsl
… servers/rendering/renderer_rd/shaders/ssao_importance_map.glsl
… servers/rendering/renderer_rd/shaders/ssao_interleave.glsl
… servers/rendering/renderer_rd/shaders/ssao_minify.glsl
… servers/rendering/rendering_server_default.h
… servers/rendering/rendering_server_wrap_mt.h
… servers/rendering_server.cpp
… servers/rendering_server.h