Supersedes: https://github.com/godotengine/godot/pull/42077

This is a direct port of Intel's [ASSAO](https://github.com/GameTechDev/ASSAO) using some optimizations from AMD's [CACAO](https://github.com/GPUOpen-Effects/FidelityFX-CACAO).

This is significantly faster and higher quality than our current SSAO. It is also much higher quality than MSSAO in #42077 but is slower. 

Timing on RX 5600 XT
ASSAO
|                   | Lowest Quality | Highest Quality|
|-------------|-----------------|------------------|
| Full Size     | 0.2 ms     |  1.1 ms    |
| Half Size    |  0.12 ms    |   0.46 ms        |

SAO (currently broken on master)
|                   | Lowest Quality | Highest Quality|
|-------------|-----------------|------------------|
| Full Size     | 2 ms    |  4ms   |
| Half Size    |      |           |

Old SAO
![sao-hq-sponza](https://user-images.githubusercontent.com/16521339/101593653-23ec2e00-39a5-11eb-8a08-602bf04dc49d.png)

New ASSAO
![ASSAO-uq-sponza](https://user-images.githubusercontent.com/16521339/101593666-2a7aa580-39a5-11eb-8809-482f6179cc64.png)

New ASSAO - Low Quality - Half Resolution
![ASSAO-lq-hr-sponza](https://user-images.githubusercontent.com/16521339/101594274-416dc780-39a6-11eb-902c-34c0340758ce.png)


<details><summary>Not used MSSAO</summary>

![mssao-hq-sponza](https://user-images.githubusercontent.com/16521339/101593781-6150bb80-39a5-11eb-838f-a7e96f63a9d6.png)

</details>