@Calinou 2020-12-09
Is it intended that I can see the textures' details on the ASSAO screenshot? I can't on the SAO and MSSAO screenshots.