@clayjohn 2020-12-09
@Calinou yep! ASSAO uses the normal buffer to generate SSAO while our implementation of SAO ignored it and MSSAO couldn't use it. 

So what your seeing is not the texture's AO, but a high detail AO generated from normals.