@Calinou 2020-12-09
@clayjohn Does this mean AO maps are still relevant? What about https://github.com/godotengine/godot/issues/43957 ?