@clayjohn 2020-12-09
@Calinou AO texture maps are much higher quality than SSAO, plus they can be used even when the object doesn't have a normal map. There is no reason not to use both AO maps and SSAO together IMO.