@reduz 2020-12-22
Look OK, I am still in doubt about exporting so many parameters to Environment, but I need to try them for a while to see what is actually important.