@Calinou 2020-12-22
> Compute shaders again or something backportable to 3.x?

Indeed, compute shaders are still required for this feature.