@clayjohn 2020-12-22
@Zireael07 @Calinou in theory the original ASSAO could be ported to GLES3 as it's all done in a pixel shader. However, it would require duplicating all the work I just did here.