@atirut-w 2021-09-09
> it would require duplicating all the work I just did here.

I think it would still be worth it to backport, though. IMO, there might be 3.x games that want realistic graphics but cannot move to 4.x for some reasons, and #48503 affects them.