@Calinou 2021-09-09
@atirut-w Someone was working on implementing GTAO in `3.x` instead: https://github.com/godotengine/godot-proposals/issues/3223